LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.create2.round.worker-expected.txt


PASS createImageData(w, h) is rounded the same as getImageData(0, 0, w, h) 
